mock_trial_ai/requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.7.1
altair==5.5.0
annotated-types==0.7.0
anthropic==0.43.0
anyio==3.7.1
asgiref==3.8.1
atomicwrites==1.4.1
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dill==0.3.8
distro==1.9.0
ecdsa==0.19.0
et_xmlfile==2.0.0
fastapi==0.68.1
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.22.0
networkx==3.3
numpy==1.26.3
openai==1.59.7
openpyxl==3.1.5
packaging==23.2
pandas==2.1.4
passlib==1.7.4
pillow==10.4.0
plotly==5.18.0
pluggy==1.5.0
propcache==0.2.1
protobuf==4.25.6
psutil==7.0.0
psycopg2-binary==2.9.10
py==1.11.0
pyarrow==18.1.0
pyasn1==0.6.1
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pytest==6.2.5
pytest-asyncio==0.18.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.4.0
scipy==1.15.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==1.4.23
starlette==0.14.2
streamlit==1.32.0
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.15.0
watchdog==6.0.0
xxhash==3.5.0
yarl==1.18.3